fix(engine): de-flake rate_limit_test + surface queue_wait clock skew
Closes #30 — RateLimiterTest.EnforcesTargetRPS used to hit httpbin.org and
assert a tight ±25% RPS window. Under Rosetta on macOS arm64 runners the
curl_multi loop overhead shifted measured RPS outside that window, breaking
the v0.3.0 release. SlowMockServer is extracted from load_strategy_test
into a shared engine/tests/mock_server.hpp, and rate_limit_test now points
at /fast on the in-process mock. Both macOS architectures run ctest again
in release.yml (testPreset restored for macos-prod-x64).

Closes #28 — queue_wait_ms = max(0, perceived - wire) silently zeroed
negative deltas in release builds; only debug builds tripped the assert.
A future regression that moved submitted_at later in the pipeline would
have flattened queue_wait in production with no signal. Keep the clamp
for sub-ms jitter, but emit a warning via vayu::utils::log_warning when
the delta is below -1ms so the signal isn't lost.
